+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -160,7 +160,7 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-06-16T09:29:00.126Z")
+    updated: Optional[datetime] = parse_datetime("2026-06-16T10:59:02.870Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")