+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KV7W75A5FZ7BV4K07JBBPMJF"
+        "01KV81C1PZFQ4PB71YA6YB6V3J"
     )
     r"""The unique identifier of the data source"""